Create README.md
# lib
A repository aiming for the benefits of library professionals. As the profession is going through major changes which involves the application of Information technology heavily, so the present and aspiring librarians should be equipped with the lastest tool and techniques inorder to create a positive impact in the libraries ans information centres.